refactor: Replace async with packet-driven connection to servers
fix: Trying to fix minor script and combat related issues

+++ gmp-server/lib/server_events.h
@@ -225,6 +225,6 @@ struct OnPlayerUnspawnForEvent {
 struct OnPlayerHitEvent {
   std::optional<std::uint64_t> attacker_id;
   std::uint64_t victim_id;
-  std::int16_t damage;
+  std::int32_t damage;
   std::uint32_t damage_type;
 };
